Book Analysis Project/requirements.txt
requests
mysql-connector-python
matplotlib
pandas
seaborn
sqlalchemy
pymysql